server/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.23
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
pydantic[email]>=2.8.0
alembic>=1.12.1
python-jose[cryptography]>=3.5.0
passlib[bcrypt]>=1.7.4
